+++ src/apps/todo/validation.js
@@ -1,6 +1,6 @@
 const Joi = require('joi');
 
-const { TodoStatus } = require('../../shared/constants');
+const { TodoStatus, UserTodoPermissions } = require('../../shared/constants');
 
 const VTodoCreateUpdateSchema = Joi.object({
     title: Joi.string().trim().min(3).max(20).required(),
@@ -10,6 +10,14 @@ const VTodoCreateUpdateSchema = Joi.object({
         .valid(...Object.values(TodoStatus)),
 });
 
+const VUserTodoPermissionSchema = Joi.object({
+    permissions: Joi.array()
+        .items(...Object.values(UserTodoPermissions))
+        .required()
+        .min(1),
+});
+
 module.exports = {
     VTodoCreateUpdateSchema,
+    VUserTodoPermissionSchema,
 };