+++ src/apps/userTodo/service.js
@@ -1,0 +1,8 @@
+const Database = require('../../shared/database');
+const UserTodoModel = require('./model');
+
+class UserTodoService extends Database {}
+
+const userTodoService = new UserTodoService(UserTodoModel);
+
+module.exports = userTodoService;