+++ src/server.js
@@ -1,28 +1,36 @@
 const express = require('express');
 const cors = require('cors');
+const swaggerJSDoc = require('swagger-jsdoc');
+const swaggerUi = require('swagger-ui-express');
 
-const middleware = require('./shared/middleware');
+// eslint-disable-next-line object-curly-newline
+const { checkAuth, checkJson, handleError, handleInvalidPath, logRequest } = require('./shared/middleware');
 const utils = require('./shared/utils');
+const { SwaggerOptions } = require('./shared/constants');
 
 const app = express();
+const specs = swaggerJSDoc(SwaggerOptions);
 
 app.use(cors());
-app.use(middleware.checkJson);
-app.use(middleware.logRequest);
+app.use(checkJson);
+app.use(logRequest);
 
 app.get(
-    '/',
+    '/async-error',
     // eslint-disable-next-line no-unused-vars
     utils.asyncWrapper(async (req, res) => {
         throw new Error('Invalid Path');
     })
 );
 
-app.use('/', require('./apps/auth/auth.routes'));
-app.use('/todo', middleware.checkAuth, require('./apps/todo/todo.routes'));
-app.use('/user', middleware.checkAuth, require('./apps/user/user.routes'));
+app.get('/', (req, res) => res.send('Ok'));
+app.use('/', require('./apps/auth/route'));
+app.use('/todos', checkAuth, require('./apps/todo/route'));
+app.use('/users', checkAuth, require('./apps/user/route'));
 
-app.use(middleware.handleInvalidPath);
-app.use(middleware.handleError);
+app.use('/api-docs', swaggerUi.serve, swaggerUi.setup(specs));
+
+app.use(handleInvalidPath);
+app.use(handleError);
 
 module.exports = { app };